// src/app/store/actions/builder.action.types.ts
export enum builderActions {
    GET_ITEMS = '[Builder Component] Get All Items',
    GET_ITEMS_SUCCESS = '[Builder API] Items Loaded Success',
    GET_DROPPED_ITEMS = '[Builder Component] Get Dropped Items',
    GET_DROPPED_ITEMS_SUCCESS = '[Builder API] Dropped Items Loaded Success',
    UPDATE_ITEM = '[Builder Component] UpdateItem',
    UPDATE_ITEM_SUCCESS = '[Builder API] Item Update Success',
    DELETE_ITEM = '[Builder Component] DeleteItem',
    DELETE_ITEM_SUCCESS = '[Builder API] Item Delete Success',
    DROP_ITEM = '[Builder Component] DropItem',
    DROP_ITEM_SUCCESS = '[Builder Component] DropItem Success',
    GET_FORM_STYLES = '[Builder Component] Get Form Styles',
    GET_FORM_STYLES_SUCCESS = '[Builder Component] Get Form Styles Success',
    UPDATE_FORM_STYLES = '[Builder Component] Update Form Styles',
    UPDATE_FORM_STYLES_SUCCESS = '[Builder Component] Update Form Styles Success',
    ERROR = '[Builder API]  Error',
    CLOSE_ERROR = '[Builder API]  Close Error'
}